[Spring 지하철 노선도 - 3단계] 오찌(오지훈) 미션 제출합니다. (#323)
* test: 테스트 시 데이터베이스 초기화를 DirtiesContext 대신 sql로 처리

* refactor: 데이터 중복, 없음 상황 처리 시 커스텀 예외를 만들도록 변경

* refactor: 서비스에서의 데이터 중복 및 존재 여부 검증 방법 변경

* refactor: toEntity 제거

* feat: 노선에 포함되는 구간을 받아서 역 목록을 돌려주는 기능 추가

* feat: 데이터베이스에 노선을 저장하는 기능 추가

* feat: 데이터베이스에서 노선이 포함하고 있는 구간들을 조회해오는 기능 추가

* feat: 데이터베이스에서 노선에 포함되면서 지정한 역을 포함하고 있는 구간을 지우는 기능 추가

* test: 서비스 레이어의 테스트를 SpringBootTest로 변경

* feat: 데이터베이스에서 동일한 노선이 존재하는지 확인하는 기능 추가

* feat: Sections에서 하행선 기준 시점 -> 종점으로 정렬된 역 목록을 반환하는 기능 추가

* refactor: 역 목록을 돌려주는 기능을 Sections에 위임

* refactor: 역 목록을 정렬하는 메서드 분리

* feat: Sections에 구간을 삽입하는 기능 추가

* feat: Sections에서 구간을 삭제하는 기능 추가

* feat: Sections끼리 비교해서 비교 대상에서 포함하지 않는 구간들만 반환하는 기능 추가

* test: 테스트들의 공통 변수 추출

* feat: 구간 생성 시 시점과 종점이 같은 역이면 예외를 반환하는 기능 추가

* refactor: 객체를 넣어서 조회하는 경우의 메서드 이름 변경

* feat: 노선 생성 요청을 받아서 노선과 구간을 저장하는 기능 추가

* feat: 구간 생성 요청을 받아서 구간을 생성하는 기능 추가

* refactor: Sections의 중복 부분 메서드 추출

* refactor: 조회 메서드에 readOnly, 삭제 메서드에 @Transactional 적용

* test: extracting을 사용해 테스트 코드 리팩토링

* refactor: 서비스 레이어 메서드 네이밍 변경

* refactor: LineService에서 StationDao 대신 StationService를 사용하도록 변경

* feat: 요청으로 들어온 노선과 역 정보로 구간을 지우는 기능 추가

* test: LineService 의 중복 변수를 전역으로 빼고 일부 테스트 Disabled 제거

* test: 단건 조회에 대한 StationService 테스트 추가

* refactor: 노선 생성과 수정 DTO 분리

* test: 테스트에 쓰이는 지하철 역 및 노선 정보 변경

* refactor: Line에서 Sections 제거

* refactor: LineResponse 생성 시 포함된 역 목록을 가지고 생성하도록 변경

* fix: 인수 테스트 실패 오류 수정

* test: 공통 테스트 코드를 AcceptanceTest로 이동

* feat: 구간 추가 및 삭제 기능 컨트롤러에 매핑

* fix: Seciton과 Line, Station 간에 외래키 추가

* feat: 해당 역을 포함하고 있는 구간이 있는 역을 삭제하려는 경우 예외를 반환하는 기능 추가

* fix: 테스트 실패 오류 수정

* refactor: 불필요한 객체 생성 부분 제거

* refactor: 메서드 명이 기능을 모두 담도록 변경

* fix: 구간 삭제 시 노선이 존재하는지 확인하도록 수정

* refactor: 끝 구간인지 확인하는 메서드 추출

* refactor: 쿼리문 가독성 좋게 개행 추가

* chore: exception 디렉토리를 더 세세하게 분리

* refactor: DataoNotFoundException을 도메인에 따라 상세하게 분리

* refactor: 노선 단 건 조회를 Optional 사용 방식으로 변경

* refactor: 역 단 건 조회를 Optional 사용 방식으로 변경

* test: 구간 추가 시 없는 역으로 생성하려는 경우 테스트 추가

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -2,14 +2,15 @@ package wooteco.subway.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
+import wooteco.subway.exception.duplicate.DuplicateStationException;
+import wooteco.subway.exception.invalidrequest.InvalidStationDeleteRequestException;
+import wooteco.subway.exception.notfound.StationNotFoundException;
 
 @Service
 public class StationService {
@@ -22,16 +23,19 @@ public class StationService {
 
     @Transactional
     public StationResponse create(StationRequest stationRequest) {
-        Station station = stationRequest.toEntity();
-        try {
-            Station newStation = stationDao.save(station);
-            return new StationResponse(newStation);
-        } catch (DuplicateKeyException e) {
-            throw new DuplicateKeyException("이미 존재하는 역입니다.");
+        Station station = new Station(stationRequest.getName());
+        validateUnique(station);
+        Station newStation = stationDao.save(station);
+        return new StationResponse(newStation);
+    }
+
+    private void validateUnique(Station station) {
+        if (stationDao.existsName(station)) {
+            throw new DuplicateStationException("이미 존재하는 역 이름입니다.");
         }
     }
 
-    public List<StationResponse> showAll() {
+    public List<StationResponse> findAll() {
         List<Station> stations = stationDao.findAll();
 
         return stations.stream()
@@ -39,16 +43,26 @@ public class StationService {
                 .collect(Collectors.toUnmodifiableList());
     }
 
+    public Station findById(Long id) {
+        return stationDao.findById(id)
+                .orElseThrow(() -> new StationNotFoundException("존재하지 않는 역입니다."));
+    }
+
     public void delete(Long id) {
         validateExist(id);
+        validateDeletable(id);
         stationDao.deleteById(id);
     }
 
     private void validateExist(Long id) {
-        try {
-            stationDao.findById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new EmptyResultDataAccessException("존재하지 않는 역입니다.", 1);
+        if (!stationDao.existsId(id)) {
+            throw new StationNotFoundException("존재하지 않는 역입니다.");
+        }
+    }
+
+    private void validateDeletable(Long id) {
+        if (stationDao.existsContainingSection(id)) {
+            throw new InvalidStationDeleteRequestException("해당 역을 포함하고 있는 구간이 있어 삭제할 수 없습니다.");
         }
     }
